requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib[bcrypt]
python-multipart
aiosmtplib
jinja2
email-validator
python-dotenv
